requirements.txt
Django==4.2.4
djangorestframework==3.14.0
python-decouple==3.8 # .env file
psycopg2==2.9.7
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
mock==5.1.0

# TODO separete dev dependencies local  and prod
pre-commit==3.3.3 # hooks for git: flake8, black, isort, bandit 

pylint-django==2.5.3
#pdb==0.13.13 # install only if neeeded
gunicorn==21.2.0
